Clean up Test Suite
