+++ phpunit.xml.dist
@@ -11,7 +11,7 @@
          stopOnFailure="false"
          failOnError="false"
          syntaxCheck="false"
-         bootstrap="./tests/bootstrap.php"
+         bootstrap="./vendor/autoload.php"
 >
 
 	<logging>
