+++ tests/Ogone/Tests/DirectLink/DirectLinkMaintenanceRequestTest.php
@@ -5,12 +5,10 @@
 
 namespace Ogone\Tests\DirectLink;
 
-use Ogone\Tests;
 use Ogone\Tests\ShaComposer\FakeShaComposer;
 use Ogone\DirectLink\DirectLinkMaintenanceRequest;
-use Ogone\DirectLink\Alias;
 
-class DirectLinkMaintenanceRequestTest extends \TestCase
+class DirectLinkMaintenanceRequestTest extends \PHPUnit_Framework_TestCase
 {
 
     /** @test */
@@ -81,5 +79,18 @@ class DirectLinkMaintenanceRequestTest extends \TestCase
             array('setAmount', '232'),
             array('setAmount', 2.32),
         );
+    }
+
+    /** @return DirectLinkMaintenanceRequest */
+    private function provideMinimalDirectLinkMaintenanceRequest()
+    {
+        $directLinkRequest = new DirectLinkMaintenanceRequest(new FakeShaComposer());
+        $directLinkRequest->setPspid('123456');
+        $directLinkRequest->setUserId('user_1234');
+        $directLinkRequest->setPassword('abracadabra');
+        $directLinkRequest->setPayId('12345678');
+        $directLinkRequest->setOperation('REN');
+
+        return $directLinkRequest;
     }
 }
