+++ tests/Ogone/Tests/DirectLink/DirectLinkQueryRequestTest.php
@@ -5,12 +5,10 @@
 
 namespace Ogone\Tests\DirectLink;
 
-use Ogone\Tests;
 use Ogone\Tests\ShaComposer\FakeShaComposer;
 use Ogone\DirectLink\DirectLinkQueryRequest;
-use Ogone\DirectLink\Alias;
 
-class DirectLinkQueryRequestTest extends \TestCase
+class DirectLinkQueryRequestTest extends \PHPUnit_Framework_TestCase
 {
 
     /** @test */
@@ -68,5 +66,17 @@ class DirectLinkQueryRequestTest extends \TestCase
             array('setPassword', '12'),
             array('setUserid', '12'),
         );
+    }
+
+    /** @return DirectLinkQueryRequest */
+    private function provideMinimalDirectLinkQueryRequest()
+    {
+        $directLinkRequest = new DirectLinkQueryRequest(new FakeShaComposer());
+        $directLinkRequest->setPspid('123456');
+        $directLinkRequest->setUserId('user_1234');
+        $directLinkRequest->setPassword('abracadabra');
+        $directLinkRequest->setPayId('12345678');
+
+        return $directLinkRequest;
     }
 }
