+++ tests/Ogone/Tests/FormGenerator/UrlGeneratorTest.php
@@ -2,12 +2,13 @@
 
 namespace Ogone\Tests\FormGenerator;
 
+use Ogone\Tests\TestCase;
 use Ogone\Ecommerce\EcommercePaymentRequest;
 use Ogone\FormGenerator\UrlGenerator;
 use Ogone\PaymentRequest;
 use Ogone\Tests\ShaComposer\FakeShaComposer;
 
-class UrlGeneratorTest extends \TestCase
+class UrlGeneratorTest extends TestCase
 {
 
     /** @test */
