--- tests/bootstrap.php
@@ -1,15 +1,0 @@
-<?php
-
-/*
- * This file is part of the Marlon Ogone package.
- *
- * (c) Marlon BVBA <info@marlon.be>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-error_reporting(-1);
-
-require_once __DIR__.'/../vendor/autoload.php';
-require_once __DIR__.'/TestCase.php';
